Users can now sign our public key.
Right now this is a bit of a hack, the user goes to /new and enters our
public key, the system checks for the finger print and then writes out the
new public key (importing into gpg and exporting first though!).  Eventually
this will be moved to a different page, and I may implement some sort of
validation scheme (i.e. require that the user validate us).