Added email spoof/forgery example.
This is an advocacy tool which will send an email with a falsified
from header to someone.  It requires that the recipient provide their
publickey, which has been signed by the service, and provide a signed
email address which will be used as the 'from' address for the example
email.

The email validation step could probably stand some more testing.
I'm not fealing too confident about it, simply because a user can
submit any data he wants into that field.

Nothing currently links to this.